image scan bean in common lib for imagescanevent
