--- common-lib/imageScan/bean.go
@@ -1,1 +1,0 @@
-package imageScan
